+++ libs/tup-components/src/allocations/index.ts
@@ -1,0 +1,1 @@
+export { AllocationsTable } from './AllocationTable';
